update Contact Header Fragments
