+++ src/components/Contacts/ContactDetails/ContactDetailsHeader/ContactHeaderSection/ContactHeaderPhone.graphql
@@ -1,0 +1,8 @@
+fragment ContactHeaderPhone on Contact {
+  id
+  primaryPerson {
+    primaryPhoneNumber {
+      number
+    }
+  }
+}
