+++ src/components/Contacts/ContactDetails/ContactDetailsHeader/ContactHeaderSection/ContactHeaderStatus.graphql
@@ -1,0 +1,4 @@
+fragment ContactHeaderStatus on Contact {
+  id
+  status
+}
